[feature/SBOT-79] 게시물 조회 API 구현

+++ common/src/commonMain/kotlin/common/model/reseponse/post/PostResponse.kt
@@ -12,7 +12,7 @@ data class PostResponse(
     @SerialName("author") val author: UserResponse,
     @SerialName("votes") val voteResponses: List<VoteResponse>,
     @SerialName("categories") val categories: List<CategoryResponse>,
-    @SerialName("selected_vote") val selectedVote: Long?,
+    @SerialName("selected_vote") val selectedVote: Int?,
     @SerialName("is_liked") val isLiked: Boolean?,
     @SerialName("created_at") val createdAt: String,
     @SerialName("updated_at") val updatedAt: String?
